requirements.txt
redis
sqlmodel
psycopg2-binary
celery
redis
python-dotenv
httpx
click
edgartools
pytest